ajusta disposição dos signals
