+++ sapl/base/apps.py
@@ -7,3 +7,6 @@ class AppConfig(django.apps.AppConfig):
     name = 'sapl.base'
     label = 'base'
     verbose_name = _('Dados Básicos')
+
+    def ready(self):
+        from sapl.base import receivers
