+++ sapl/materia/apps.py
@@ -6,6 +6,3 @@ class AppConfig(apps.AppConfig):
     name = 'sapl.materia'
     label = 'materia'
     verbose_name = _('Matéria')
-
-    def ready(self):
-        from sapl.base import receivers
